refactor: Post, Channel로 컴포넌트 이름 변경
