+++ next-app/src/components/common/Skeleton/SubscriptionType.tsx
@@ -2,7 +2,7 @@ import Skeleton from 'react-loading-skeleton'
 
 import Flex from '../Flex'
 import Anchor from '../Anchor'
-import { Container } from '../FeedItem/Subscription/SubscriptionType.style'
+import { Container } from '../FeedItem/Channel/Channel.style'
 
 const SubscriptionType = () => {
   return (
